requirements.txt
backcall==0.2.0
beautifulsoup4==4.9.3
dataclasses==0.6
decorator==4.4.2
html5lib==1.1
ipython==7.18.1
ipython-genutils==0.2.0
jedi==0.17.2
logzero==1.5.0
lxml==4.5.2
numpy==1.19.2
pandas==1.1.3
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.0
pkg-resources==0.0.0
prompt-toolkit==3.0.8
ptyprocess==0.6.0
Pygments==2.7.1
python-dateutil==2.8.1
pytz==2020.1
screeninfo==0.6.5
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
tabulate==0.8.7
traitlets==5.0.5
urllib3==1.25.10
wcwidth==0.2.5
webencodings==0.5.1
